+++ esbuild.js
@@ -34,7 +34,13 @@ async function main() {
 		sourcesContent: false,
 		platform: 'node',
 		outfile: 'dist/extension.js',
-		external: ['vscode', 'node-window-manager', 'extract-file-icon'],
+		external: [
+			'vscode',
+			'node-window-manager',
+			'extract-file-icon',
+			'jsdom',
+			'open',
+		],
 		logLevel: 'silent',
 		plugins: [
 			copy({